Fix TypeScript definitions and export all interfaces
Fixes #85; required a little bit of creative thinking to be able to export both the module and the interfaces; this isn't very nice as it results in consumers having to know the layout of the repo in order to import the interfaces; however it does the job and I doubt js-logger will see any extensive refactors in the foreseeable future :)

Add a test to ensure we don't break typescript definitions in the future whilst I was at it.
