+++ package.json
@@ -22,7 +22,9 @@
   "main": "src/logger.js",
   "scripts": {
     "lint": "jshint --exclude=src/*.min.js src/*.js test-src/*.js",
-    "test": "node-qunit-phantomjs ./test-src/index.html",
+    "test": "npm run test:unit && npm run test:tsd",
+    "test:unit": "node-qunit-phantomjs ./test-src/index.html",
+    "test:tsd": "cd test-src/typescript-consumer && tsc --noEmit index",
     "build": "npm run build:version && npm run build:minify",
     "build:minify": "uglifyjs src/logger.js --mangle --compress -o src/logger.min.js",
     "build:version": "cross-env replace -s 'VERSION = \"[^\"]+\"' \"VERSION = \\\"$npm_package_version\\\"\" src/logger.js"
@@ -33,6 +35,7 @@
     "node-qunit-phantomjs": "^2.0.0",
     "replace": "^1.0.0",
     "rimraf": "^2.6.2",
+    "typescript": "^3.1.3",
     "uglify-js": "^3.4.9"
   }
 }
